+++ DLSFTPClient/DemoApp/FileDownloadViewController.m
@@ -202,7 +202,7 @@
 
     DLSFTPClientFailureBlock failureBlock = ^(NSError *error) {
         dispatch_async(dispatch_get_main_queue(), ^{
-            NSString *errorString = [NSString stringWithFormat:@"Error %d", error.code];
+            NSString *errorString = [NSString stringWithFormat:@"Error %ld", (long)error.code];
             UIAlertView *alertView = [[UIAlertView alloc] initWithTitle:errorString
                                                                 message:error.localizedDescription
                                                                delegate:nil
